requirements.txt
streamlit>=1.20.0
opencv-python>=4.7.0.72
numpy>=1.23.0
pillow>=9.0.0
pandas>=1.4.0
